+++ readium/shared/r2-shared/src/main/java/org/readium/r2/shared/Experimental.kt
@@ -14,7 +14,7 @@ package org.readium.r2.shared
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.TYPEALIAS)
 annotation class PdfSupport
 
-@RequiresOptIn(message = "Usage of Fragment in the Navigator is still experimental. The API may be changed in the future without notice.")
+@RequiresOptIn(message = "Support for SearchService is still experimental. The API may be changed in the future without notice.")
 @Retention(AnnotationRetention.BINARY)
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
-annotation class FragmentNavigator
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.TYPEALIAS, AnnotationTarget.PROPERTY)
+annotation class Search