+++ readium/shared/r2-shared/src/test/java/org/readium/r2/shared/publication/LocatorTest.kt
@@ -295,3 +295,213 @@ class LocatorTest {
     }
 
 }
+
+class LocatorCollectionTest {
+
+    @Test fun `parse {LocatorCollection} minimal JSON`() {
+        assertEquals(
+            LocatorCollection(),
+            LocatorCollection.fromJSON(JSONObject("{}"))
+        )
+    }
+
+    @Test fun `parse {LocatorCollection} full JSON`() {
+        assertEquals(
+            LocatorCollection(
+                metadata = LocatorCollection.Metadata(
+                    localizedTitle = LocalizedString.fromStrings(mapOf(
+                        "en" to "Searching <riddle> in Alice in Wonderlands - Page 1",
+                        "fr" to "Recherche <riddle> dans Alice in Wonderlands – Page 1"
+                    )),
+                    numberOfItems = 3,
+                    otherMetadata = mapOf(
+                        "extraMetadata" to "value"
+                    )
+                ),
+                links = listOf(
+                    Link(rels = setOf("self"), href = "/978-1503222687/search?query=apple", type = "application/vnd.readium.locators+json"),
+                    Link(rels = setOf("next"), href = "/978-1503222687/search?query=apple&page=2", type = "application/vnd.readium.locators+json"),
+                ),
+                locators = listOf(
+                    Locator(
+                        href = "/978-1503222687/chap7.html",
+                        type = "application/xhtml+xml",
+                        locations = Locator.Locations(
+                            fragments = listOf(":~:text=riddle,-yet%3F'"),
+                            progression = 0.43
+                        ),
+                        text = Locator.Text(
+                            before = "'Have you guessed the ",
+                            highlight = "riddle",
+                            after = " yet?' the Hatter said, turning to Alice again."
+                        )
+                    ),
+                    Locator(
+                        href = "/978-1503222687/chap7.html",
+                        type = "application/xhtml+xml",
+                        locations = Locator.Locations(
+                            fragments = listOf(":~:text=in%20asking-,riddles"),
+                            progression = 0.47
+                        ),
+                        text = Locator.Text(
+                            before = "I'm glad they've begun asking ",
+                            highlight = "riddles",
+                            after = ".--I believe I can guess that,"
+                        )
+                    )
+                )
+            ),
+            LocatorCollection.fromJSON(JSONObject("""{
+              "metadata": {
+                "title": {
+                    "en": "Searching <riddle> in Alice in Wonderlands - Page 1",
+                    "fr": "Recherche <riddle> dans Alice in Wonderlands – Page 1"
+                },
+                "numberOfItems": 3,
+                "extraMetadata": "value"
+              },
+              "links": [
+                {"rel": "self", "href": "/978-1503222687/search?query=apple", "type": "application/vnd.readium.locators+json"},
+                {"rel": "next", "href": "/978-1503222687/search?query=apple&page=2", "type": "application/vnd.readium.locators+json"}
+              ],
+              "locators": [
+                {
+                  "href": "/978-1503222687/chap7.html",
+                  "type": "application/xhtml+xml",
+                  "locations": {
+                    "fragments": [
+                      ":~:text=riddle,-yet%3F'"
+                    ],
+                    "progression": 0.43
+                  },
+                  "text": {
+                    "before": "'Have you guessed the ",
+                    "highlight": "riddle",
+                    "after": " yet?' the Hatter said, turning to Alice again."
+                  }
+                },
+                {
+                  "href": "/978-1503222687/chap7.html",
+                  "type": "application/xhtml+xml",
+                  "locations": {
+                    "fragments": [
+                      ":~:text=in%20asking-,riddles"
+                    ],
+                    "progression": 0.47
+                  },
+                  "text": {
+                    "before": "I'm glad they've begun asking ",
+                    "highlight": "riddles",
+                    "after": ".--I believe I can guess that,"
+                  }
+                }
+              ]
+            }"""))
+        )
+    }
+
+    @Test fun `get {Locator} minimal JSON`() {
+        assertJSONEquals(
+            JSONObject("""{
+                "locators": []
+            }"""),
+            LocatorCollection().toJSON()
+        )
+    }
+
+    @Test fun `get {Locator} full JSON`() {
+        assertJSONEquals(
+            JSONObject("""{
+              "metadata": {
+                "title": {
+                    "en": "Searching <riddle> in Alice in Wonderlands - Page 1",
+                    "fr": "Recherche <riddle> dans Alice in Wonderlands – Page 1"
+                },
+                "numberOfItems": 3,
+                "extraMetadata": "value"
+              },
+              "links": [
+                {"rel": ["self"], "href": "/978-1503222687/search?query=apple", "type": "application/vnd.readium.locators+json", "templated": false},
+                {"rel": ["next"], "href": "/978-1503222687/search?query=apple&page=2", "type": "application/vnd.readium.locators+json", "templated": false}
+              ],
+              "locators": [
+                {
+                  "href": "/978-1503222687/chap7.html",
+                  "type": "application/xhtml+xml",
+                  "locations": {
+                    "fragments": [
+                      ":~:text=riddle,-yet%3F'"
+                    ],
+                    "progression": 0.43
+                  },
+                  "text": {
+                    "before": "'Have you guessed the ",
+                    "highlight": "riddle",
+                    "after": " yet?' the Hatter said, turning to Alice again."
+                  }
+                },
+                {
+                  "href": "/978-1503222687/chap7.html",
+                  "type": "application/xhtml+xml",
+                  "locations": {
+                    "fragments": [
+                      ":~:text=in%20asking-,riddles"
+                    ],
+                    "progression": 0.47
+                  },
+                  "text": {
+                    "before": "I'm glad they've begun asking ",
+                    "highlight": "riddles",
+                    "after": ".--I believe I can guess that,"
+                  }
+                }
+              ]
+            }"""),
+            LocatorCollection(
+                metadata = LocatorCollection.Metadata(
+                    localizedTitle = LocalizedString.fromStrings(mapOf(
+                        "en" to "Searching <riddle> in Alice in Wonderlands - Page 1",
+                        "fr" to "Recherche <riddle> dans Alice in Wonderlands – Page 1"
+                    )),
+                    numberOfItems = 3,
+                    otherMetadata = mapOf(
+                        "extraMetadata" to "value"
+                    )
+                ),
+                links = listOf(
+                    Link(rels = setOf("self"), href = "/978-1503222687/search?query=apple", type = "application/vnd.readium.locators+json"),
+                    Link(rels = setOf("next"), href = "/978-1503222687/search?query=apple&page=2", type = "application/vnd.readium.locators+json"),
+                ),
+                locators = listOf(
+                    Locator(
+                        href = "/978-1503222687/chap7.html",
+                        type = "application/xhtml+xml",
+                        locations = Locator.Locations(
+                            fragments = listOf(":~:text=riddle,-yet%3F'"),
+                            progression = 0.43
+                        ),
+                        text = Locator.Text(
+                            before = "'Have you guessed the ",
+                            highlight = "riddle",
+                            after = " yet?' the Hatter said, turning to Alice again."
+                        )
+                    ),
+                    Locator(
+                        href = "/978-1503222687/chap7.html",
+                        type = "application/xhtml+xml",
+                        locations = Locator.Locations(
+                            fragments = listOf(":~:text=in%20asking-,riddles"),
+                            progression = 0.47
+                        ),
+                        text = Locator.Text(
+                            before = "I'm glad they've begun asking ",
+                            highlight = "riddles",
+                            after = ".--I believe I can guess that,"
+                        )
+                    )
+                )
+            ).toJSON()
+        )
+    }
+
+}